+++ mcomix/archive/archive_base.py
@@ -11,7 +11,8 @@ from mcomix import portability
 from mcomix import i18n
 from mcomix import process
 from mcomix import callback
-from mcomix import archive
+from mcomix.archive import password as archive_password
+
 
 class BaseArchive(object):
     """ Base archive interface. All filenames passed from and into archives
@@ -120,7 +121,7 @@ class BaseArchive(object):
         If an empty string is set, assume that the user did not provide
         a password. """
 
-        password = archive.password.ask_for_password(self.archive)
+        password = archive_password.ask_for_password(self.archive)
         if password is None:
             password = ""